// src/helpers/types/types.ts
export type TIngredientData = {
  calories: number;
  carbohydrates: number;
  fat: number;
  image: string;
  image_large: string;
  image_mobile: string;
  name: string;
  price: number;
  proteins: number;
  type: string;
  __v: number;
  _id: string;
}

export type TIngredient = {
    data: TIngredientData[];
    success: boolean;
}